browser_automation/requirements.txt
browser-use
langchain-openai
playwright
streamlit
python-dotenv
